+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatListUI.java
@@ -106,7 +106,7 @@ public class FlatListUI
 			@Override
 			public void focusLost( FocusEvent e ) {
 				super.focusLost( e );
-				toggleSelectionColors( false );
+				toggleSelectionColors( e.isTemporary() );
 			}
 		};
 	}